% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.createRandomizedOverlapsFile.r
\name{ch.createRandomizedOverlapsFile}
\alias{ch.createRandomizedOverlapsFile}
\title{A function to randomize the relation between the probe names and the overlaps. This is used to test a Null model.}
\usage{
ch.createRandomizedOverlapsFile(data, probeCols, shuffleNum = 10)
}
\arguments{
\item{data}{A dataframe with the overlap data processed from a typical values experiment.}

\item{probeCols}{a vector of strings that specifies the names of the columns in "data" that contains the the probes.}

\item{shuffleNum}{an integer specifying the number of times to "shuffle" the vector when randomizing the order. DEFAULT = 10.}
}
\value{
the overlaps dataframe with the newly randomized relations between the overlaps and the prompt names. WARNING: Do Not Confuse this file with the orginal file that has the correct relation between prompts and overlaps.
}
\description{
A function to randomize the relation between the probe names and the overlaps. This is used to test a Null model.
}
\examples{
ch.createRandomizedOverlapsFile (df.myInputData, c("Item1", "Item2"))
}
\keyword{overlap}
\keyword{prompts}
\keyword{randomize}
