% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.filterOverlapsByQuantile.r
\name{ch.filterOverlapsByQuantile}
\alias{ch.filterOverlapsByQuantile}
\title{A function to filters the choice data by value overlaps using quantiles rather than fixed values.}
\usage{
ch.filterOverlapsByQuantile(
  data,
  snCol,
  RTcol,
  overlapRoundCol,
  correctCol,
  correctVals = c(TRUE, FALSE),
  chanceThreshold = 1,
  lowRTquantileThreshold = 0,
  highRTquantileThreshold = 1,
  minOverlapN = 0,
  statsOutputFile = NULL
)
}
\arguments{
\item{data}{A dataframe containing the choice RT data. (typically after running through ch.moralsDataPrep()).}

\item{snCol}{A string that specifies the name of the column in "data" that contains the subject number.}

\item{RTcol}{A string that specifies the name of the column in "data" that contains the RT for each trial.}

\item{overlapRoundCol}{A string that specifies the name of the column in "data" that contains the rounded overlaps for the item in each trial. Column created in ch.moralsDataPrep() via ch.mergeChoiceDataWithOverlapsData().}

\item{correctCol}{A string that specifies the name of the new column that will contains an indication of whether or not the response was correct.}

\item{correctVals}{A vector of two values that specifies the "correct" value (index 1) and the "incorrect" value (index 2). e.g, c("yes", "no")}

\item{chanceThreshold}{A number specifying the quantile that all subjects whose p(HVO) falls below it will be removed from the dataset. DEFAULT = 0 (none removed)}

\item{lowRTquantileThreshold}{A number specifying the quantile that, for each overlapRound, individual RTs that fall below it will be removed from the dataset. DEFAULT = 0.0 (none removed)}

\item{highRTquantileThreshold}{A number specifying the quantile that, for each overlapRound, individual RTs that fall above  it will be removed from the dataset. DEFAULT = 1.0 (none removed)}

\item{minOverlapN}{A number specifying the minimum number of responses an overlapRound condition must have to remain in the dataset. DEFAULT = 0 (none removed)}

\item{statsOutputFile}{the filename that you want the statistics summary output written to. DEFAULT = NULL (no file written)}
}
\value{
a dataframe of filtered data.
}
\description{
This function filters the choice data for analysis by removing RT outliers and bad subjects based on RT and p(HVO) using quantiles rather than fixed values.
}
\examples{
ch.filterOverlapsByQuantile (data=moralsData, "sn", "RT", "overlapRound", "avePred", correctVals = c(TRUE, FALSE), chanceThreshold = 0.5, lowRTquantileThreshold = 0.025, highRTquantileThreshold = 0.975, minOverlapN = 40)
}
\keyword{choice}
\keyword{data}
\keyword{filter}
\keyword{overlap}
