% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getMoralsItemColumnNames.r
\name{ch.getMoralsItemColumnNames}
\alias{ch.getMoralsItemColumnNames}
\title{A function to create a vector of the column names containing the items in the morals experiment}
\usage{
ch.getMoralsItemColumnNames(params, colBaseName = "Item")
}
\arguments{
\item{params}{a list of parameters that are read in using "ch.readMoralsDBfile.r."}

\item{colBaseName}{a string with the base of the column name containing the items. DEfAULT = "Item"}
}
\value{
the vector of string names
}
\description{
This function creates a vector of string names with the colBaseName followed by numbers to specify the column names
}
\examples{
ch.getMoralsItemColumnNames (params)
}
\keyword{column}
\keyword{item}
\keyword{morals}
\keyword{names}
