% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getMoralsItemsInOverlap.r
\name{ch.getMoralsItemsInAllOverlaps}
\alias{ch.getMoralsItemsInAllOverlaps}
\title{A function to get an ordered, non-duplicate list of the combinations of Items for each overlap}
\usage{
ch.getMoralsItemsInAllOverlaps(data, overlapCol, itemCols)
}
\arguments{
\item{data}{a dataframe.}

\item{overlapCol}{a string specifying the name of the column in the dataframe containing the overlap variable.}

\item{itemCols}{a vector containing strings that specify the names of the columns in the dataframe containing the item variables.}
}
\value{
a list containing: 1) a dataframe containing an ordered, non-duplicate combinations of Items for each overlap, and 2) a dataframe containing the number of non-duplicate combinations of items for each overlap.
}
\description{
This function outputs a list containing: 1) a dataframe containing an ordered, non-duplicate combinations of Items for each overlap, and 2) a dataframe containing the number of non-duplicate combinations of items for each overlap.
}
\examples{
ch.getItemsInAllOverlaps (df, "overlapRound", c("Item1", "Item2"))
}
