% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getMoralsRTaxisName.r
\name{ch.getMoralsRTaxisName}
\alias{ch.getMoralsRTaxisName}
\title{A function to get the RT axis label}
\usage{
ch.getMoralsRTaxisName(transformType, useResids)
}
\arguments{
\item{transformType}{a string that is either "log" (transform RT to logs) or something else.}

\item{useResids}{boolean that specifies whether we use the residuals to a learning function as the DV.}
}
\value{
a string that specifies the RT axis label.
}
\description{
This function gets the RT axis label.
}
\examples{
ch.getMoralsRTaxisName ("log", TRUE)
}
\keyword{RT}
\keyword{axis}
\keyword{label}
\keyword{morals}
