% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.globalFilterByQuantile.r
\name{ch.globalFilterByQuantile}
\alias{ch.globalFilterByQuantile}
\title{A function to filter RT data based on the coefficient of variation.}
\usage{
ch.globalFilterByQuantile(
  data,
  snCol,
  RTcol,
  CVlowQuantileThreshold = 0,
  CVhighQuantileThreshold = 1,
  statsOutputFile = NULL
)
}
\arguments{
\item{data}{A dataframe containing the choice RT data. (typically after running through ch.moralsDataPrep()).}

\item{snCol}{a string that specifies the name of the column in "data" that contains the subject number.}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each trial.}

\item{CVlowQuantileThreshold}{A number specifying the quantile that all subjects whose coefficient of variation falls below it will be removed from the dataset. DEFAULT = 0 (none removed)}

\item{CVhighQuantileThreshold}{A number specifying the quantile that all subjects whose coefficient of variation falls above it will be removed from the dataset. DEFAULT = 1 (none removed)}

\item{statsOutputFile}{the filename that you want the statistic summary output written to. DEFAULT = NULL (no file written)}
}
\value{
a dataframe of filtered data.
}
\description{
This function filters RT data based on the coefficient of variation.  Specifically, it removes subjects based on their variability of RTs relative to the subjects in the dataset (by quantile).
}
\examples{
ch.globalFilterByQuantile (data=moralsData, "sn", "RT", CVlowQuantileThreshold = 0, CVhighQuantileThreshold = 0.95)
}
\keyword{filter}
\keyword{global}
\keyword{quantile}
