% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.mergeChoiceDataWithOverlapsData.r
\name{ch.mergeChoiceDataWithOverlapsData}
\alias{ch.mergeChoiceDataWithOverlapsData}
\title{A function to prepare the morals data for analysis}
\usage{
ch.mergeChoiceDataWithOverlapsData(
  data,
  overlap.data,
  overlapCol,
  directionCol,
  respChoiceCol,
  respChoiceVal = c("Item1", "Item2"),
  item1cols = c("Item1"),
  item2cols = c("Item2"),
  overlapItem1cols = c("IA1"),
  overlapItem2cols = c("IB1"),
  roundThreshold = 0.1,
  roundDirection = ceiling,
  numOverlapBins = 10,
  outfile = NULL
)
}
\arguments{
\item{data}{A dataframe with the choice data from a typical PVT experiment.}

\item{overlap.data}{A dataframe with the overlap data processed from a typical values experiment.}

\item{overlapCol}{a string that specifies the name of the column in "data" that contains the overlaps for the item in each trial.}

\item{directionCol}{a string that specifies the name of the column in "data" that contains the direction of the overlap for each trial.}

\item{respChoiceCol}{a string that specifies the name of the column in "data" that contains the the participant's response to the prompt - yes take action or no take no action.}

\item{respChoiceVal}{a vector of two values that specifies the choose Item1 option ("yes" take action in many morals experimants, thus saving Item1) value (index 1) and the choose Item1 option ("no" take no action in many morals experimants, thus saving Item2) value (index 2).}

\item{item1cols}{a vector of strings that specifies the names of the columns in "data" that contains the the probes in Item 1.}

\item{item2cols}{a vector of strings that specifies the names of the columns in "data" that contains the the probes in Item 2.}

\item{overlapItem1cols}{a vector of strings that specifies the names of the columns in the overlaps data (overlap.data) that contains the correspoinding probes in Item 1.}

\item{overlapItem2cols}{a vector of strings that specifies the names of the columns in the overlaps data (overlap.data) that contains the correspoinding probes in Item 2.}

\item{roundThreshold}{An integer that specifies the nearest interval that the overlaps should be rounded to. DEFAULT = 0.1 (round to the nearest 0.1)}

\item{roundDirection}{An option that specifies the rounding direction: ceiling (always round up), floor (always round down), or round (round to the nearest value, up or down). DEFAULT = ceiling}

\item{numOverlapBins}{An integer that specifies the number of bins the overlaps should be binned into. This is only used if roundThreshold = NULL.  DEFAULT = 10 (round overlaps so that there are 10 bins)}

\item{outputFile}{the filename that you want the output written to. DEFAULT = NULL (no file written)}

\item{params}{a list of parameters that are read in using "ch.readMoralsDBfile.r."}
}
\value{
a dataframe of merged data.
}
\description{
This function prepares the morals data for analysis by creating overlapRound, correct, and a few other columns.  It also makes sure the trials starts at 1, and computes average p(hit) and RT for each subject.
}
\examples{
ch.mergeChoiceDataWithOverlapsData (data=moralsData, "sn", "RT", "overlap", "direction", "trials", "respDef", respChoiceVal = c("Yes", "No"))
}
\keyword{choice}
\keyword{datafile}
\keyword{merge}
\keyword{morals}
\keyword{overlaps}
