% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsAnalyzeLearningEffect.r
\name{ch.moralsAnalyzeLearningEffect}
\alias{ch.moralsAnalyzeLearningEffect}
\title{A function to analyze learning effects}
\usage{
ch.moralsAnalyzeLearningEffect(data, trialCol, RTCol, fitCol, resCol, params,
  filenameID = "gp")
}
\arguments{
\item{data}{the morals dataframe after running through ch.moralsDataPrep().}

\item{trialCol}{a string that specifies the name of the column in "data" that contains the trial number.}

\item{fitCol}{a string that specifies the name of the new column that will contain the predicted datapoints.}

\item{resCol}{a string that specifies the name of the new column that will contain the residual datapoints.}

\item{params}{a list of parameters that are read in using "ch.readMoralsDBfile.r."}

\item{filenameID}{a string that will be added to the file names and titles that makes them unique if you run this call several times.}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each trial.}
}
\value{
a list containing: data = original datafram with the new columns; nlsFit = the nls fit object.
}
\description{
This function attempts to fit a non-linear, decelerating function to the specified data; adds the predicted and residual datapoints to the dataframe; and plots the learning function.
}
\examples{
ch.moralsAnalyzeLearningEffect (data=moralsData,"trial", "RT", "fitRT", "resRT", params)
}
\keyword{analysis}
\keyword{learning}
\keyword{morals}
