% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsDataClean.r
\name{ch.moralsDataClean}
\alias{ch.moralsDataClean}
\title{A function to cleans the choice data for analysis}
\usage{
ch.moralsDataClean(
  data,
  snCol,
  RTcol,
  trialCol,
  respChoiceCol,
  respChoiceVal = c("Item1", "Item2"),
  dropVarFilename = NULL,
  dropSNFilename = NULL,
  RTlogTransform = FALSE,
  outfile = NULL
)
}
\arguments{
\item{data}{A dataframe containing the choice data.}

\item{snCol}{a string that specifies the name of the column in "data" that contains the subject number.}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each trial.}

\item{trialCol}{a string that specifies the name of the column in "data" that contains the trial number.}

\item{respChoiceCol}{a string that specifies the name of the column in "data" that contains the the participant's response to the prompt - yes take action or no take no action.}

\item{respChoiceVal}{a vector of two values that specifies the choose Item1 option ("yes" take action in many morals experimants, thus saving Item1) value (index 1) and the choose Item1 option ("no" take no action in many morals experimants, thus saving Item2) value (index 2).}

\item{dropVarFilename}{A filename of a file that contains a vector of column names to be removed from "data" DEFAULT = NULL (remove nothing)}

\item{dropSNFilename}{A filename of a file that contains a vector of subject names to be removed from "data" DEFAULT = NULL (remove nothing)}

\item{RTlogTransform}{A boolean to specify whether to transform the RT variable to log(RT). DEFAULT = FALSE}

\item{outputFile}{the filename that you want the output written to. DEFAULT = NULL (no file written)}
}
\value{
a dataframe of cleaned data.
}
\description{
This function cleans the choice data for analysis by removing bad subjects (based on sn column), dropping columns, making sure the trials starts at 1, and computing average p(HVO) and RT for each subject.
}
\examples{
ch.moralsDataClean (data=moralsData, "sn", "RT", "trials", "keyDef", dropVarFilename ="dropVars.txt", dropSNFilename = "badSNs.txt")
}
\keyword{choice}
\keyword{clean}
\keyword{data}
