% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsDataPrep.r
\name{ch.moralsDataPrep}
\alias{ch.moralsDataPrep}
\title{A function to prepare the morals data for analysis}
\usage{
ch.moralsDataPrep(
  data,
  snCol,
  RTcol,
  overlapCol,
  directionCol,
  trialCol,
  respChoiceCol,
  respChoiceVal = c("Item1", "Item2"),
  item1cols = c("Item1"),
  item2cols = c("Item2"),
  overlapItem1cols = c("IA1"),
  overlapItem2cols = c("IB1"),
  params
)
}
\arguments{
\item{data}{morals dataframe.}

\item{snCol}{a string that specifies the name of the column in "data" that contains the subject number.}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each trial.}

\item{overlapCol}{a string that specifies the name of the column in "data" that contains the overlaps for the item in each trial.}

\item{directionCol}{a string that specifies the name of the column in "data" that contains the direction of the overlap for each trial.}

\item{trialCol}{a string that specifies the name of the column in "data" that contains the trial number.}

\item{respChoiceCol}{a string that specifies the name of the column in "data" that contains the the participant's response to the prompt - yes take action or no take no action.}

\item{respChoiceVal}{a vector of two values that specifies the choose Item1 option ("yes" take action in many morals experimants, thus saving Item1) value (index 1) and the choose Item1 option ("no" take no action in many morals experimants, thus saving Item2) value (index 2).}

\item{item1cols}{a vector of strings that specifies the names of the columns in "data" that contains the the probes in Item 1.}

\item{item2cols}{a vector of strings that specifies the names of the columns in "data" that contains the the probes in Item 2.}

\item{overlapItem1cols}{a vector of strings that specifies the names of the columns in the overlaps file that contains the correspoinding probes in Item 1.}

\item{overlapItem2cols}{a vector of strings that specifies the names of the columns in the overlaps file that contains the correspoinding probes in Item 2.}

\item{params}{a list of parameters that are read in using "ch.readMoralsDBfile.r."}
}
\value{
a dataframe of prepared data.  It also writes the data to prepDataOutFile (specified in params) which will be used by other functions.
}
\description{
This function prepares the morals data for analysis by creating overlapRound, correct, and a few other columns.  It also makes sure the trials starts at 1, and computes average p(hit) and RT for each subject.
}
\examples{
ch.moralsDataPrep (data=moralsData, "sn", "RT", "overlap", "direction", "trials", "respDef", respChoiceVal = c("Yes", "No"), params=parameters)
}
\keyword{data}
\keyword{morals}
\keyword{prep}
