% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsDprimeAnalysis.r
\name{ch.moralsDprimeAnalysis}
\alias{ch.moralsDprimeAnalysis}
\title{A function to calculate a dPrime analysis on the Morals data.}
\usage{
ch.moralsDprimeAnalysis(
  data,
  overlapRoundCol,
  correctCol,
  correctVals,
  targetPresentCol,
  targetPresentVals,
  resCol,
  addCorrection = TRUE,
  params,
  printR2 = TRUE,
  filenameID = NULL
)
}
\arguments{
\item{data}{the morals dataframe after running through ch.moralsDataPrep().}

\item{overlapRoundCol}{a string that specifies the name of the column in "data" that contains the overlap round column.}

\item{correctCol}{a string the specifies the column name in "data" that contains the variable whether the participant's response was correct.}

\item{correctVals}{a vector of two values that specifies the "correct" value (index 1) and the "incorrect" value (index 2). e.g, c("yes", "no")}

\item{targetPresentCol}{a string that specifies the name of the column in "data" that contains the whether or not the target was presented.}

\item{targetPresentVals}{a vector of two values that specifies the "target present" value (index 1) and the "target absent" value (index 2). e.g, c("yes", "no")}

\item{resCol}{a string that specifies the name of the new column that will contain the residual datapoints.}

\item{addCorrection}{a boolean that specifies whether you want a .5 correction to be added the total hits, FAs, misses, and CRs. This corrects for 0 and 1 values for FA and Hits. DEFAULT = T.}

\item{params}{a list of parameters that are read in using "ch.readMoralsDBfile.r."}
}
\value{
a dataframe with the dPrime tatistics.
}
\description{
This function analyzes the group  morals data.
}
\examples{
ch.moralsDprimeAnalysis (data=moralsData,"overlapRound", "correct", c(1,0), "targetPresent", c(TRUE,FALSE), params = params)
}
\keyword{analysis}
\keyword{d}
\keyword{dPrime}
\keyword{data}
\keyword{morals}
\keyword{prime}
