% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ch.moralsEquateItemStringsForNumberQuantDataset.r
\name{ch.moralsEquateItemStringsForNumberQuantDataset}
\alias{ch.moralsEquateItemStringsForNumberQuantDataset}
\title{A function to equate the probe 1 and probe 2 strings when quantity varies.}
\usage{
ch.moralsEquateItemStringsForNumberQuantDataset(data, probe1Col, probe2Col)
}
\arguments{
\item{data}{morals dataframe.}

\item{probe1Col}{a string the specifies the column name in "data" that contains the first probe item.}

\item{probe2Col}{a string the specifies the column name in "data" that contains the second probe item.}
}
\value{
the dataframe (data) with the four new columns: "p1Quant" and "p1Quant" (the quantities of probe 1 and 2 respectively) & "p1" and "p2" (the singular, equalized names of probe 1 and 2 respoectively).
}
\description{
This function equates the probe 1 and probe 2 strings when quantity varies.
}
\examples{
ch.moralsEquateItemStringsForNumberQuantDataset (analysisReadyData.gp, "probe1", "probe2")
}
\keyword{equate}
\keyword{morals}
\keyword{probes}
\keyword{quantity}
