% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsGetAndPlotQuantDirOverlap.r
\name{ch.moralsGetAndPlotQuantDirOverlap}
\alias{ch.moralsGetAndPlotQuantDirOverlap}
\title{A function to get and plot directional overlaps by quantity.}
\usage{
ch.moralsGetAndPlotQuantDirOverlap(
  data,
  probe1Col,
  probe2Col,
  overlapCol,
  directionCol,
  quantValueCuts,
  filename = NULL,
  ...
)
}
\arguments{
\item{data}{a dataframe containing overlap data for probe combinations.}

\item{probe1Col}{a string the specifies the column name in "data" that contains the first probe item.}

\item{probe2Col}{a string the specifies the column name in "data" that contains the second probe item.}

\item{overlapCol}{a string that specifies the name of the column in "data" that contains the overlap column.}

\item{directionCol}{a string that specifies the name of the column in "data" that contains the direction column that indications the direction of the overlap.}

\item{quantValueCuts}{a vector of numbers that specifies the cutoff values for each quantity.  The first category is <= to the first cutoff value; each successive category (c1, c2, etc) is c1 < X <= c2; the final category is greater than the last number in the vector.}

\item{filename}{the filename (pdf) to save the figure.  DEFAULT = NULL (file not saved)}
}
\value{
a dataframe with the probe in one column, it's average directional overlap in another, its N in another, followed by separate columns containing the average directional overlap and N for each group.
}
\description{
This function gets and plots directional overlaps by quantity. It simply combines four other functions: ch.moralsGetDirOverlapForAllProbes(); ch.moralsQuantsToGrps(); ch.moralsGetDirOverlapByGrp(); and ch.stripPlotByGrp().
}
\examples{
ch.moralsGetDirOverlapForAllProbes (overlapData, "IA1", "IB1", "overlap","direction")
}
\keyword{directional}
\keyword{morals}
\keyword{overlaps}
\keyword{plot}
\keyword{quantity}
