% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsGetDirOverlapForAllProbes.r
\name{ch.moralsGetDirOverlapForAllProbes}
\alias{ch.moralsGetDirOverlapForAllProbes}
\title{A function to calculates the directional overlap for all probes.}
\usage{
ch.moralsGetDirOverlapForAllProbes(data, probe1Col, probe2Col, overlapCol,
  directionCol)
}
\arguments{
\item{data}{a dataframe containing overlap data for probe combinations.}

\item{probe1Col}{a string the specifies the column name in "data" that contains the first probe item.}

\item{probe2Col}{a string the specifies the column name in "data" that contains the second probe item.}

\item{overlapCol}{a string that specifies the name of the column in "data" that contains the overlap column.}

\item{directionCol}{a string that specifies the name of the column in "data" that contains the direction column that indications the direction of the overlap.}
}
\value{
the dataframe (data), doubled in size because it adds the reverse order of the probes, with the directional overlap column (dirOverlap).
}
\description{
This function calculates the directional overlap for all probes and ensures every combination is present in both directions (e.g., itemA vs itemB and itemB vs itemA). Thus, there will be duplicate combinations with oppoisite directional overlaps.
}
\examples{
ch.moralsGetDirOverlapForAllProbes (overlapData, "IA1", "IB1", "overlap","direction")
}
