% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsGetProbNo.r
\name{ch.moralsGetProbNo}
\alias{ch.moralsGetProbNo}
\title{A function to assess the participants bias to say no as a function of "x"}
\usage{
ch.moralsGetProbNo(
  data,
  x,
  yesNoCol,
  yesNoVal = c("Yes", "No"),
  summarize = T,
  minNperXbin = 0,
  plotFilename = NULL,
  ...
)
}
\arguments{
\item{data}{the dataframe.}

\item{x}{a string the specifies the column name in "data" that contains the x variable.}

\item{yesNoCol}{a string the specifies the column name in "data" that contains the variable with the participant's yes/no response.}

\item{summarize}{Do you want to collapse the data by "x"? DEFAULT = TRUE.}

\item{minNperXbin}{an integer that specifies the minimum number of trials necessary to include an group of the x variable in the graph. DEFAULT = 0.}

\item{plotFilename}{the filename to save the plot (pdf) if you want it. DEFAULT = NULL (no plot saved).}

\item{respYNVal}{a vector of two values that specifies the yes "take action" value (index 1) and the no "take no action" value (index 2). e.g, c("yes", "no")}
}
\value{
an lm fit object.
}
\description{
This function assesses the participants bias to say no as a function of "x"
}
\examples{
ch.moralsGetProbNo (data=moralsData, "overlapRound", "respDef", c("Yes", "No"))
}
\keyword{morals}
\keyword{no}
\keyword{probability}
