% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsGrpRTpHit.r
\name{ch.moralsGrpRTpHit}
\alias{ch.moralsGrpRTpHit}
\title{A function to analyze the group Morals data}
\usage{
ch.moralsGrpRTpHit(data, trialCol, RTCol, fitCol, resCol, overlapRoundCol,
  yesNoCol, yesNoVal = c("Yes", "No"), correctCol,
  correctVals = c(TRUE, FALSE), params)
}
\arguments{
\item{data}{the morals dataframe after running through ch.moralsDataPrep().}

\item{trialCol}{a string that specifies the name of the column in "data" that contains the trial number.}

\item{fitCol}{a string that specifies the name of the new column that will contain the predicted datapoints.}

\item{resCol}{a string that specifies the name of the new column that will contain the residual datapoints.}

\item{overlapRoundCol}{a string that specifies the name of the column in "data" that contains the overlap column.}

\item{yesNoCol}{a string the specifies the column name in "data" that contains the variable with the participant's yes/no response.}

\item{yesNoVal}{a vector of two values that specifies the yes "take action" value (index 1) and the no "take no action" value (index 2). e.g, c("yes", "no")}

\item{correctCol}{a string that specifies the name of the new column that will contains a "1" if the participant chose the item with the greatest value distribution and a "0" if they did not.}

\item{correctVals}{a vector of two values that specifies the "correct" value (index 1) and the "incorrect" value (index 2). e.g, c("yes", "no")}

\item{params}{a list of parameters that are read in using "ch.readMoralsDBfile.r."}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each trial.}
}
\value{
dataframe with the learning function fit and residuals
}
\description{
This function analyzes the group  morals data.
}
\examples{
ch.moralsGrpRTpHit (data=moralsData,"trial", "RT", "res.RT", "fit.RT", "overlap", "keyDef", c("Yes", "No"), "correct", c("yes", "no"), params=parameters)
}
\keyword{analysis}
\keyword{data}
\keyword{group}
\keyword{morals}
