% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsItemRTpHit.r
\name{ch.moralsItemRTpHit}
\alias{ch.moralsItemRTpHit}
\title{A function to analyze the Morals data by subject}
\usage{
ch.moralsItemRTpHit(
  data,
  itemCols,
  resCol,
  overlapRoundCol,
  correctCol,
  correctVals = c(TRUE, FALSE),
  useTwoParameterModel = FALSE,
  minNperOverlap = 0,
  minUniqueOverlaps = 3,
  statsOutputFile = NULL,
  numPlotRows = 2,
  dt.set = NULL
)
}
\arguments{
\item{data}{the morals dataframe after running through ch.moralsDataPrep().}

\item{itemCols}{a vector of strings that specifies the names of the columns in "data" that contains the the probes. This is all the columns for all of the groups}

\item{resCol}{a string that specifies the name of the new column that will contain the residual datapoints.}

\item{overlapRoundCol}{a string that specifies the name of the column in "data" that contains the overlap column.}

\item{correctVals}{a vector of two values that specifies the "correct" value (index 1) and the "incorrect" value (index 2). e.g, c("yes", "no")}

\item{useTwoParameterModel}{A boolean that specifies whether to use a two parameter p(HOV) model.  If this is set to TRUE, then this function will fit a p(HVO) model whereby the rightmost point (overlap = 1.0) is not fixed at p(HVO) = 0.5. DEFAULT = FALSE.}

\item{minNperOverlap}{an integer that specifies the minimum number of trials necessary to include an overlap bin in the graph. DEFAULT = 0.}

\item{minUniqueOverlaps}{An integer specifying the minimum number of unique overlap bins necessary for the program to calculate the pHVO and RT function.  DEFAULT = 3.}

\item{statsOutputFile}{the filename that you want the statistics summary output written to. DEFAULT = NULL (construct filename from "params")}

\item{numPlotRows}{an integer that specifies the number of rows in the output figure, DEFAULT = 2)}

\item{dt.set}{A string that specifies the a short title to identify the output of these plots. DEFAULT = NULL)}
}
\value{
dataframe with the learning function fit and residuals
}
\description{
This function analyzes the group  morals data.
}
\examples{
ch.moralsSnRTpHit (data=moralsData,"sn", "trial", "RT", "res.RT", "fit.RT", "overlap", "keyDef", c("Yes", "No"), "correct", params=parameters)
}
\keyword{analysis}
\keyword{by}
\keyword{data}
\keyword{morals}
\keyword{subject}
