% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsPlotLearningEffect.r
\name{ch.moralsPlotLearningEffect}
\alias{ch.moralsPlotLearningEffect}
\title{A function to plot learning effects}
\usage{
ch.moralsPlotLearningEffect(
  data,
  trialCol,
  RTCol,
  fitCol,
  resCol,
  yLabel,
  filenameID = "gp",
  ...
)
}
\arguments{
\item{data}{the morals dataframe after running through ch.moralsDataPrep().}

\item{trialCol}{a string that specifies the name of the column in "data" that contains the trial number.}

\item{fitCol}{a string that specifies the name of the new column that will contain the predicted datapoints.}

\item{resCol}{a string that specifies the name of the new column that will contain the residual datapoints.}

\item{yLabel}{a string that specifies the yLabel for the graphs}

\item{filenameID}{a string that will be added to the file names and titles that makes them unique if you run this call several times.}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each trial.}
}
\value{
.
}
\description{
This function plots the learning function.
}
\examples{
ch.moralsPlotLearningEffects (data=moralsData,"trial", "RT", "fitRT", "resRT", RTlabel)
}
\keyword{learning}
\keyword{morals}
\keyword{plots}
