% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsPlotSnRTpHitFits.r
\name{ch.moralsPlotSnRTpHitFits}
\alias{ch.moralsPlotSnRTpHitFits}
\title{A function to plot the subject fits}
\usage{
ch.moralsPlotSnRTpHitFits(
  snSummaryArray,
  snCol,
  rtSloCol,
  rtIntCol,
  rtR2Col,
  phBCol,
  phACol,
  phR2Col,
  xAll,
  filename = NULL
)
}
\arguments{
\item{snSummaryArray}{a dataframe containing the morals fit coeficients and r square for RT and p(Hit) for each subject.}

\item{snCol}{a string that specifies the name of the column in "snSummaryArray" that contains the subject number.}

\item{rtSloCol}{a string that specifies the name of the column in "snSummaryArray" that contains the best fit slope for the RT data of each subject.}

\item{rtIntCol}{a string that specifies the name of the column in "snSummaryArray" that contains the best fit intercept for the RT data of each subject.}

\item{rtR2Col}{a string that specifies the name of the column in "snSummaryArray" that contains the r square the fit of the RT data of each subject.}

\item{phBCol}{a string that specifies the name of the column in "snSummaryArray" that contains the best fit beta for the p(hit) data of each subject.}

\item{phACol}{a string that specifies the name of the column in "snSummaryArray" that contains the best fit alpha for the p(hit) data of each subject.}

\item{phR2Col}{a string that specifies the name of the column in "snSummaryArray" that contains the r square the fit of the p(hit) data of each subject.}

\item{xAll}{a vector containing the overlap rounds used when fitting the RT and p(hit) data.}

\item{filename}{a string with the filename (pdf) for the plot to be saved. DEFAULT = NULL (no file saved)}
}
\value{
.
}
\description{
This function plots the subject fits.
}
\examples{
ch.moralsPlotSnRTpHitFits(subOutData, "sn", "rtSlo", "rtInt", "rtR2", "phB", "phA", "phR2", xAll, filename)
}
\keyword{fit}
\keyword{morals}
\keyword{plots}
\keyword{subject}
