% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsQuantsToGrps.r
\name{ch.moralsQuantsToGrps}
\alias{ch.moralsQuantsToGrps}
\title{A function to equate the probe 1 and probe 2 strings when quantity varies and bin the quantities into categories.}
\usage{
ch.moralsQuantsToGrps(data, probe1Col, probe2Col, quantValueCuts)
}
\arguments{
\item{data}{morals dataframe.}

\item{probe1Col}{a string the specifies the column name in "data" that contains the first probe item.}

\item{probe2Col}{a string the specifies the column name in "data" that contains the second probe item.}

\item{quantValueCuts}{a vector of numbers that specifies the cutoff values for each quantity.  The first category is <= to the first cutoff value; each successive category (c1, c2, etc) is c1 < x <= c2; the final category is greater than the last number in the vector.}
}
\value{
the dataframe (data) with six new columns: "p1Quant" and "p1Quant" (the quantities of probe 1 and 2 respectively); "p1" and "p2" (the singular, equalized names of probe 1 and 2 respoectively); and "p1GrpSize" and "p2GrpSize" (the group categories based on quantValueCuts for probe 1 and 2 respectively).
}
\description{
This function equates the probe 1 and probe 2 strings when quantity varies and bins the quantities into categories.
}
\examples{
ch.moralsQuantsToGrps (analysisReadyData.gp, "probe1", "probe2", c(1, 10, 40))
}
\keyword{equate}
\keyword{group}
\keyword{morals}
\keyword{quantity}
