% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsRTpHitFit.r
\name{ch.moralsRTpHitFit}
\alias{ch.moralsRTpHitFit}
\title{A function to fit the RT and pHit function to morals data}
\usage{
ch.moralsRTpHitFit(data, overlapRoundCol, RTCol, correctCol,
  correctVals = c(TRUE, FALSE), cex1 = 1.5, cex.topTile = 1.25,
  printR2 = F, topTitle = NULL, filename = NULL, ...)
}
\arguments{
\item{data}{the morals dataframe after running through ch.moralsDataPrep().}

\item{overlapRoundCol}{a string that specifies the name of the column in "data" that contains the overlap column.}

\item{correctCol}{a string that specifies the name of the new column that will contains a "1" if the participant chose the item with the greatest value distribution and a "0" if they did not.}

\item{correctVals}{a vector of two values that specifies the "correct" value (index 1) and the "incorrect" value (index 2). e.g, c("yes", "no")}

\item{cex1}{sets the default font size. DEFAULT = 1.25.}

\item{cex.topTile}{sets the default font size of the title at the top of the page. DEFAULT = 1.25.}

\item{printR2}{a boolean that determines whether to print the r square on the graph. DEFAULT = FALSE.}

\item{topTitle}{a string that will be the title at the top of the page. DEFAULT = NULL.}

\item{filename}{the filename (pdf) to save the output of the graph. DEFAULT = NULL (no graph saved)}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each trial.}
}
\value{
a list containing: RTfit = lm object with the fit of the RT data; pHitFit = the nls fit object with the fit of the pHit data; pHitR2 = the r2 of the pHitFit.
}
\description{
This function attempts to fit a non-linear, decelerating function to the pHit data; a linear function to the RT data;  plots the  functions on a single graph; and returns a list with the fit statistics.
}
\examples{
ch.moralsRTpHitFit (data=moralsData,"overlapRound", "resdRT", "correct", c("yes", "no"), filename = "plot.pdf")
}
\keyword{RT}
\keyword{fit}
\keyword{morals}
\keyword{pHit}
\keyword{plots}
