% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsSnRTpHit.r
\name{ch.moralsSnRTpHit}
\alias{ch.moralsSnRTpHit}
\title{A function to analyze the Morals data by subject}
\usage{
ch.moralsSnRTpHit(
  data,
  snCol,
  trialCol,
  RTCol,
  fitCol,
  resCol,
  overlapCol,
  correctCol,
  correctVals = c(TRUE, FALSE),
  useTwoParameterModel = FALSE,
  params
)
}
\arguments{
\item{data}{the morals dataframe after running through ch.moralsDataPrep().}

\item{snCol}{a string that specifies the name of the column in "data" that contains the subject number.}

\item{trialCol}{a string that specifies the name of the column in "data" that contains the trial number.}

\item{fitCol}{a string that specifies the name of the new column that will contain the predicted datapoints.}

\item{resCol}{a string that specifies the name of the new column that will contain the residual datapoints.}

\item{overlapCol}{a string that specifies the name of the column in "data" that contains the overlap column.}

\item{correctVals}{a vector of two values that specifies the "correct" value (index 1) and the "incorrect" value (index 2). e.g, c("yes", "no")}

\item{useTwoParameterModel}{A boolean that specifies whether to use a two parameter p(HOV) model.  If this is set to TRUE, then this function will fit a p(HVO) model whereby the rightmost point (overlap = 1.0) is not fixed at p(HVO) = 0.5. DEFAULT = FALSE.}

\item{params}{a list of parameters that are read in using "ch.readMoralsDBfile.r."}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the RT for each trial.}
}
\value{
dataframe with the learning function fit and residuals
}
\description{
This function analyzes the group  morals data.
}
\examples{
ch.moralsSnRTpHit (data=moralsData,"sn", "trial", "RT", "res.RT", "fit.RT", "overlap", "keyDef", c("Yes", "No"), "correct", params=parameters)
}
\keyword{analysis}
\keyword{by}
\keyword{data}
\keyword{morals}
\keyword{subject}
