% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.moralsXXYPlotFitsByGrp.r
\name{ch.moralsXXYPlotFitsByGrp}
\alias{ch.moralsXXYPlotFitsByGrp}
\title{A function to plot multiple model fits on one graph}
\usage{
ch.moralsXXYPlotFitsByGrp(df.models, grp1Col, grp2Col = NULL, xCol, data,
  subsetCond = NULL, subsetString = NULL, addLegend = T,
  filenameID = NULL)
}
\arguments{
\item{df.models}{a models list containing: a dfIndex, created by ch.subsetDFbyGroups() and a series of model fits output by ch.getLmModel() or ch.getPhitModel();}

\item{grp1Col}{a string that specifies the name of the column in "dfIndex" that contains one grouping variable.  This is required.}

\item{grp2Col}{a string that specifies the name of the column in "dfIndex" that contains another grouping variable.  This is optional.  A max of two grouping variables is allowed}

\item{xCol}{a string that specifies the name of the column in "data" that contains the predictor (x) variables for the fit specified in the models.}

\item{data}{the morals dataframe after running through ch.moralsDataPrep().}

\item{subsetCond}{an expression that is used to subset df.models$dfIndex to include only the group conditions that you want to graph (e.g., cond1 == "pretest" & sn == "4").}

\item{subsetString}{an string that mirrors the subsetCond (e.g., "cond1 == pretest & sn == 4"). This is used for titling the graph.}

\item{addLegend}{a boolean that specifies whether a legend should be added to the graphs. DEFAULT = T}

\item{filenameID}{a string that will be put on all filenames when the plots are saved. DEFAULT = NULL (no file saved)}
}
\value{
.
}
\description{
This function plots multiple model fits on one graph. The model fits should have one DV, one predictor variable (xCol), and one or two grouping variables (grp1Col & grp2Col)
}
\examples{
ch.moralsXXYPlotFitsByGrp(df.models, "scenarioType", "agent", "overlapRound", moralsData)
}
\keyword{fit}
\keyword{group}
\keyword{model}
\keyword{morals}
\keyword{plots}
