% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.readMoralsDBfile.r
\name{ch.readMoralsDBfile}
\alias{ch.readMoralsDBfile}
\title{A function to read the morals dbfile}
\usage{
ch.readMoralsDBfile(dbfile = "moralsDBfile.txt")
}
\arguments{
\item{dbfile}{the filename of the dbfile.}
}
\value{
a list with all the parameterd
}
\description{
This function reads the dbfile for the morals analysis.  The dbfile contains key words and their values for the analysis.
}
\examples{
ch.readMoralsDBfile ("myDBfile.txt")
}
\keyword{dbfile}
\keyword{morals}
\keyword{read}
