% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.rmLearningBySN.r
\name{ch.rmLearningBySN}
\alias{ch.rmLearningBySN}
\title{A function remove the learning effect on RT by Subject}
\usage{
ch.rmLearningBySN(
  data,
  snCol,
  trialCol,
  RTcol,
  fitCol,
  resCol,
  outputFile = NULL
)
}
\arguments{
\item{data}{A dataframe containing the choice data (often after running through ch.moralsDataPrep()).}

\item{snCol}{A string that specifies the name of the column in "data" that contains the subject number.}

\item{trialCol}{A string that specifies the name of the column in "data" that contains the trial number.}

\item{RTcol}{A string that specifies the name of the column in "data" that contains the RT for each trial.}

\item{fitCol}{A string that specifies the name of the new column that will contain the predicted datapoints.}

\item{resCol}{A string that specifies the name of the new column that will contain the residual datapoints.}

\item{outputFileTag}{A string to put on the front of a filename that you want a graph of the learning fit written to. DEFAULT = NULL (no file written)}
}
\value{
A dataframe with the learning function fit and residuals included
}
\description{
This function remove the learning effect on RT by Subject.
}
\examples{
ch.rmLearningBySN (data=moralsData,"sn", "trial", "RT", "res.RT", "fit.RT", "fitOut.pdf")
}
\keyword{RT}
\keyword{by}
\keyword{learning}
\keyword{remove}
\keyword{subject}
